+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - [UIMARCAUTH-88](https://issues.folio.org/browse/UIMARCAUTH-88) Browse authorities: Add Heading type facet.
 - [UIMARCAUTH-64](https://issues.folio.org/browse/UIMARCAUTH-64) Add DELETE action for individual MARC Authority records.
 - [UIMARCAUTH-119](https://issues.folio.org/browse/UIMARCAUTH-119) Browse "Enter" run search when placeholder "Select a browse option" active
+- [UIMARCAUTH-118](https://issues.folio.org/browse/UIMARCAUTH-118) Highlighting the row with "Heading/reference" value at result list after editing of record
+
 
 ## [1.0.5](https://github.com/folio-org/ui-marc-authorities/tree/v1.0.5) (2022-04-15)